requirements.txt
# Python dependencies for Azure DevOps pipeline and automation
pyyaml>=6.0
python-dotenv>=1.0.0

